+++ src/app/PageWrapper.tsx
@@ -11,13 +11,20 @@ import { Modals } from '@/components/modals';
 import { NetworkModeToast } from '@/components/network-mode-toast';
 import { Notice } from '@/components/notice';
 import { SearchComponent } from '@/features/search/search';
-import { Indicator, StatusBar, StatusBarBase } from '@/features/status-bar/status-bar';
+import {
+  AlertBarBase,
+  Indicator,
+  StatusBar,
+  StatusBarBase,
+} from '@/features/status-bar/status-bar';
 import { Box } from '@/ui/Box';
 import { Flex } from '@/ui/Flex';
-import { Text } from '@/ui/Text';
-import { TextLink } from '@/ui/TextLink';
 import { useColorMode } from '@chakra-ui/react';
-import { FC } from 'react';
+import React, { FC } from 'react';
+import { Text } from '@/ui/Text';
+import { UnorderedList } from '@/ui/UnorderedList';
+import { ListItem } from '@/ui/ListItem';
+import { TextLink } from '@/ui/TextLink';
 
 export const PageWrapper: FC = ({ children }) => {
   const colorMode = useColorMode().colorMode;
@@ -32,6 +39,63 @@ export const PageWrapper: FC = ({ children }) => {
         backdropFilter={'blur(10px)'}
       >
         <StatusBar />
+        <AlertBarBase
+          indicator={Indicator.major}
+          content={
+            <Flex direction={'column'} gap={'8px'} flexGrow={1}>
+              <Text fontWeight={700} fontSize={'14px'} lineHeight={'16px'} color={'#303030'}>
+                Stacks network upgrade
+              </Text>
+              <UnorderedList
+                paddingLeft={'8px'}
+                fontWeight={400}
+                fontSize={'14px'}
+                lineHeight={'20px'}
+                color={'#303030'}
+              >
+                <ListItem>
+                  The Stacks network will be undergoing two hard-forks, as{' '}
+                  <TextLink
+                    display="inline"
+                    href="https://github.com/stacksgov/sips/blob/main/sips/sip-022/sip-022-emergency-pox-fix.md"
+                    target="_blank"
+                    color={'#1068BF'}
+                    borderBottom={'1px solid #1068BF'}
+                  >
+                    outlined in SIP-022
+                  </TextLink>
+                  .
+                </ListItem>
+                <ListItem>
+                  This update will require miners and operators to upgrade to{' '}
+                  <TextLink
+                    display="inline"
+                    href="https://github.com/stacks-network/stacks-blockchain/releases/tag/2.2.0.0.1"
+                    target="_blank"
+                    color={'#1068BF'}
+                    borderBottom={'1px solid #1068BF'}
+                  >
+                    stacks-node version 2.2.0.0.1
+                  </TextLink>
+                </ListItem>
+                <ListItem>
+                  Additionally, this update will disabled Stacking, and any locked STX will be
+                  unlocked. For more details,{' '}
+                  <TextLink
+                    display="inline"
+                    href="https://forum.stacks.org/t/a-bug-in-stacks-increase-call-is-impacting-stacking-rewards-this-cycle/14867"
+                    target="_blank"
+                    color={'#1068BF'}
+                    borderBottom={'1px solid #1068BF'}
+                  >
+                    see this post
+                  </TextLink>
+                  .
+                </ListItem>
+              </UnorderedList>
+            </Flex>
+          }
+        />
       </Box>
       <Flex
         maxWidth="100vw"